+++ CHANGELOG.md
@@ -21,7 +21,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 - The process list collection system info collector to now be a post-breach action. #1697
 - The "/api/monkey/download" endpoint to accept an OS and return a file. #1675
 - Log messages to contain human-readable thread names. #1766
-- The log file name to `infection-monkey-agent-<random_string>-<timestamp>.log`. #1761
+- The log file name to `infection-monkey-agent-<TIMESTAMP>-<RANDOM_STRING>.log`. #1761
 
 ### Removed
 - VSFTPD exploiter. #1533